+++ graphlearn/transform.py
@@ -5,7 +5,7 @@ little fixing to be able to be used by graphlearn.
 graphtransformer: graph -> object the decomposer unterstands how to use
 '''
 
-from eden.graph import Vectorizer
+from eden.graph  import   _edge_to_vertex_transform
 
 class GraphTransformer(object):
 
@@ -50,7 +50,7 @@ class GraphTransformer(object):
         -------
         graphwrapper : iterator
         '''
-        return [Vectorizer._edge_to_vertex_transform(i) for i in inputs]
+        return [_edge_to_vertex_transform(i) for i in inputs]
