+++ graphlearn/utils/draw.py
@@ -182,8 +182,8 @@ def set_ids(graph):
 def graphlearn(graphs,
                size=6,
                font_size=15,
-               node_size=200,
-               node_border=False,
+               #node_size=200,
+               #node_border=False,
                show_direction=False,
                abstract_color=None,
                edge_color=None,
@@ -243,8 +243,8 @@ def graphlearn(graphs,
 
     draw_graph_set(graphs,
                    size=size,
-                   node_size=node_size,
-                   node_border=node_border,
+                   #node_size=node_size,
+                   #node_border=node_border,
                    font_size=font_size,
                    edge_color=edge_color,
                    vertex_color=vertex_color,
